# Optimize the display of task process bar