--- a/setup.cfg
+++ b/setup.cfg
@@ -27,7 +27,7 @@ include_package_data = True
 packages = find:
 install_requires =
     numpy>=1.14.0
-    pandas>=1.0.0
+    pandas>=1.1.0,<=1.4.4
     scipy>=1.0.0
     scikit-learn>=0.20
     numexpr>=2.6.4